Add debian directory and script so that xmlrpc++ can be built as a full source and binary debian package.
This uses debuild (dpkg-buildpackage), instead of
dpkg-deb, which is the non-standard quick hack.

This is a (hopefully) temporary workaround to support
building NIDAS for DSMs that have been upgraded to
Debian8 armel.

Once all the DSMs are upgraded to Debian8, we can
update NIDAS to libxmlrpc-c, which is packaged
for RedHat and Debian systems. The we won't have to
maintain this old package.